# HW2/src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: Book Author API
  description: API for managing books and authors
  version: 1.0.0
servers:
  - url: localhost:8080/
paths:
  /book:
    get:
      summary: Get all books
      operationId: getAllBooks
      tags:
        - Book
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '500':
          description: Internal server error
    post:
      summary: Add a new book
      operationId: addBook
      tags:
        - Book
      requestBody:
        description: Book object to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Bad request
        '500':
          description: Internal server error
    put:
      summary: Update a book
      operationId: updateBook
      tags:
        - Book
      requestBody:
        description: Book object to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /book/{authorId}:
    get:
      summary: Get books by author
      operationId: getBooksByAuthor
      tags:
        - Book
      parameters:
        - name: authorId
          in: path
          description: ID of author to fetch
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /book/{bookId}:
    delete:
      summary: delete book by id
      operationId: deleteBook
      tags:
        - Book
      parameters:
        - name: bookId
          in: path
          description: ID of boook
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
        '404':
          description: Not found
  /author:
    get:
      summary: Get all authors
      operationId: getAllAuthors
      tags:
        - Author
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        '500':
          description: Internal server error
    put:
      summary: Update an author
      operationId: updateAuthor
      tags:
        - Author
      requestBody:
        description: Author object to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '200':
          description: Book added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /author/{bookId}:
    get:
      summary: Get authors by book
      operationId: getAuthorsByBook
      tags:
        - Author
      parameters:
        - name: bookId
          in: path
          description: ID of book
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '500':
          description: Internal server error
  /author/{authorName}:
    get:
      summary: Get authors by name
      operationId: getAuthorsByName
      tags:
        - Author
      parameters:
        - name: authorName
          in: path
          description: author name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '500':
          description: Internal server error
  /author/{authorId}:
    delete:
      summary: delete author by id
      operationId: deleteAuthor
      tags:
        - Author
      parameters:
        - name: authorId
          in: path
          description: ID of author to fetch
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
        '404':
          description: Not found
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
    Author:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string